+++ conftest.py
@@ -10,10 +10,12 @@ from wandb.apis import InternalApi
 import six
 import json
 import sys
+import time
 import threading
 import logging
 from multiprocessing import Process
 from vcr.request import Request
+import requests
 from wandb import wandb_socket
 from wandb import env
 from wandb import util
@@ -27,7 +29,10 @@ def pytest_runtest_setup(item):
     wandb.uninit()
     global_settings = os.path.expanduser("~/.config/wandb/settings")
     if os.path.exists(global_settings):
-        os.remove(global_settings)
+        try:
+            os.remove(global_settings)
+        except OSError:
+            pass
     # This is used to find tests that are leaking outside of tmp directories
     os.environ["WANDB_DESCRIPTION"] = item.parent.name + "#" + item.name
 
@@ -134,12 +139,24 @@ def wandb_init_run(request, tmpdir, request_mocker, mock_server, monkeypatch, mo
                             class Hook(object):
                                 def register(self, what, where):
                                     pass
+
+                            class Pub(object):
+                                def publish(self, **kwargs):
+                                    pass
+
+                            class Hist(object):
+                                def get_range(self, **kwargs):
+                                    return [[None, 1, ('#source code', None)]]
+
                             self.events = Hook()
+                            self.display_pub = Pub()
+                            self.history_manager = Hist()
 
                         def register_magics(self, magic):
                             pass
                     return Jupyter()
                 wandb.get_ipython = fake_ipython
+                wandb.jupyter.get_ipython = fake_ipython
             # no i/o wrapping - it breaks pytest
             os.environ['WANDB_MODE'] = 'clirun'
 
@@ -457,6 +474,15 @@ def live_mock_server(request):
     app = create_app()
     server = Process(target=app.run, kwargs={"port": port, "debug": True, "use_reloader": False})
     server.start()
+    for i in range(5):
+        try:
+            time.sleep(1)
+            res = requests.get("http://localhost:%s/storage" % port, timeout=1)
+            if res.status_code == 200:
+                break
+            print("Attempting to connect but got: %s", res)
+        except requests.exceptions.RequestException:
+            print("timed out")
     yield server
     server.terminate()
     server.join()