+++ wandb/apis/internal.py
@@ -325,6 +325,7 @@ class Api(object):
         query Viewer{
             viewer {
                 id
+                flags
                 entity
                 teams {
                     edges {
@@ -1041,7 +1042,7 @@ class Api(object):
         def no_retry_4xx(e):
             if not isinstance(e, requests.HTTPError):
                 return True
-            if not(e.response.status_code >= 400 and e.response.status_code < 500):
+            if not(e.response.status_code >= 400 and e.response.status_code < 500) or e.response.status_code == 429:
                 return True
             body = json.loads(e.response.content)
             raise UsageError(body['errors'][0]['message'])
@@ -1147,7 +1148,7 @@ class Api(object):
         def no_retry_4xx(e):
             if not isinstance(e, requests.HTTPError):
                 return True
-            if not(e.response.status_code >= 400 and e.response.status_code < 500):
+            if not(e.response.status_code >= 400 and e.response.status_code < 500) or e.response.status_code == 429:
                 return True
             body = json.loads(e.response.content)
             raise UsageError(body['errors'][0]['message'])