+++ wandb/catboost/__init__.py
@@ -1,49 +1,38 @@
-'''W&B callback for lightgbm
+'''W&B plot_feature_importances for catboost
 
-Really simple callback to get logging for each tree
+Simple function to send catboost feature importances metrics to the server
 
 Example usage:
 
-param_list = [("eta", 0.08), ("max_depth", 6), ("subsample", 0.8), ("colsample_bytree", 0.8), ("alpha", 8), ("num_class", 10)]
-config.update(dict(param_list))
-lgb = lgb.train(param_list, d_train, callbacks=[wandb_callback()])
+catboost_model.fit(...)
+.
+.
+.
+wandb.catboost.plot_feature_importances(catboost_model, X_train.columns)
 '''
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import numpy as np
-from .utils import test_types
-
-import lightgbm
 import wandb
 
+from .utils import test_fitted, test_types
+from warnings import simplefilter
 
-def wandb_callback():
-    def callback(env):
-        eval_results = {}
-        recorder = lightgbm.record_evaluation(eval_results)
-        recorder(env)
-
-        for validation_key in eval_results.keys():
-            for key in eval_results[validation_key].keys():
-                wandb.log({
-                    validation_key + "_" + key: eval_results[validation_key][key][0]
-                }, commit=False)
-        # Previous log statements use commit=False. This commits them.
-        wandb.log({})
-    return callback
+# ignore all future warnings
+simplefilter(action='ignore', category=FutureWarning)
 
 
 def plot_feature_importances(model=None, feature_names=None,
                              title='Feature Importance', max_num_features=50):
     """
-    Evaluates & plots the importance of each feature for the gbm fitting tasks.
+    Evaluates & plots the importance of each feature for the classification/regressor tasks.
 
-    Should only be called with a fitted model (otherwise an error is thrown).
-    Only works with LightGBM model that have a `feature_importance` attribute.
+    Should only be called with a fitted classifer or regressor (otherwise an error is thrown).
+    Only works with classifiers that have a feature_importances_ attribute.
 
     Arguments:
-        model: Takes in a fitted gbm i.e. LightGBM
+        model (clf): Takes in a fitted classifier or regressor.
         feature_names (list): Names for features. Makes plots easier to read by
                                 replacing feature indexes with corresponding names.
 
@@ -52,18 +41,18 @@ def plot_feature_importances(model=None, feature_names=None,
               under 'auto visualizations'.
 
     Example:
-        wandb.lightgbm.plot_feature_importances(model, ['width', 'height, 'length'])
+        wandb.catboost.plot_feature_importances(model, ['width', 'height, 'length'])
     """
-    attributes_to_check = 'feature_importance'
+    attributes_to_check = 'feature_importances_'
 
     if not hasattr(model, attributes_to_check):
         wandb.termwarn(
             "%s attribute not in model. Cannot plot feature importances." % attributes_to_check)
         return
 
-    if (test_types(model=model)):
+    if (test_types(model=model) and test_fitted(model)):
         feature_names = np.asarray(feature_names)
-        importances = model.feature_importance()
+        importances = model.feature_importances_
 
         indices = np.argsort(importances)[::-1]
 
@@ -91,5 +80,5 @@ def plot_feature_importances(model=None, feature_names=None,
                 ))
 
         wandb.log({'feature_importances': feature_importances_table(feature_names, importances)})
-        print("Go to your dashboard to see the LightGBM's Feature Importances graph plotted.")
+        print("Go to your dashboard to see the Catboost's Feature Importances graph plotted.")
         return feature_importances_table(feature_names, importances)